+++ R/svBase-package.R
@@ -18,13 +18,13 @@
 
 ## usethis namespace: start
 #' @importFrom data.table as.data.table haskey key rbindlist setattr setDT setkeyv
-#' @importFrom data.trame as.data.trame data.trame
+#' @importFrom data.trame as.data.trame data.trame is.data.trame let_data.trame_to_data.table let_data.table_to_data.trame
 #' @importFrom tibble as_tibble tibble tribble
 #' @importFrom svMisc assign_temp eval_data_dot
-#' @importFrom dplyr across add_count add_tally arrange bind_cols collect count distinct filter full_join group_by inner_join left_join mutate pull rename rename_with right_join select summarise tally transmute ungroup
+#' @importFrom dplyr across add_count add_tally arrange bind_cols collect count distinct filter full_join group_by inner_join is_grouped_df left_join mutate pull rename rename_with right_join select summarise tally transmute ungroup
 #' @importFrom tidyr drop_na extract fill pivot_longer pivot_wider replace_na separate separate_rows uncount unite
-#' @importFrom collapse fcount fnobs fselect fsummarise fmutate frename fsubset fgroup_by fgroup_vars fungroup replace_NA ss
-#' @importFrom rlang abort caller_env check_dots_empty0 eval_bare f_lhs f_rhs inject is_formula warn
+#' @importFrom collapse allNA fcount fnobs fselect fsummarise fmutate frename fsubset fgroup_vars fungroup get_collapse group_by_vars replace_NA ss fgroup_by
+#' @importFrom rlang abort caller_env check_dots_empty0 eval_bare f_name f_lhs f_rhs inject is_formula warn
 # @importFrom pillar tbl_sum
 #' @import zeallot
 ## usethis namespace: end